refactor arena setup
